requirements.txt
python-dateutil
python-dotenv
pandas
requests
